+++ src/Parser/DeathKnight/Blood/CombatLogParser.js
@@ -3,6 +3,8 @@ import HealingDone from 'Parser/Core/Modules/HealingDone';
 import DamageDone from 'Parser/Core/Modules/DamageDone';
 import DamageTaken from 'Parser/Core/Modules/DamageTaken';
 
+import DeathRecapTracker from 'Main/DeathRecapTracker';
+
 import Abilities from './Modules/Abilities';
 import Channeling from './Modules/Core/Channeling';
 import GlobalCooldown from './Modules/Core/GlobalCooldown';
@@ -22,8 +24,6 @@ import DancingRuneWeapon from './Modules/Features/DancingRuneWeapon';
 import InitialMarrowrendCast from './Modules/Features/InitialMarrowrendCast';
 import DeathStrikeTiming from './Modules/Features/DeathStrikeTiming';
 import BoneShieldTimesByStacks from './Modules/Features/BoneShieldTimesByStacks';
-import DeathRecap from './Modules/Features/DeathRecap';
-import DeathRecapTracker from './Modules/Features/DeathRecapTracker';
 
 // Resources
 import RunicPowerDetails from './Modules/RunicPower/RunicPowerDetails';
@@ -83,7 +83,6 @@ class CombatLogParser extends CoreCombatLogParser {
     souldrinker: Souldrinker,
     boneShield: BoneShield,
     boneShieldTimesByStacks: BoneShieldTimesByStacks,
-    deathRecap: DeathRecap,
     deathRecapTracker: DeathRecapTracker,
 
     // DOT
