+++ src/Parser/DeathKnight/Blood/Modules/Features/BloodDeathRecapTracker.js
@@ -1,0 +1,14 @@
+import DeathRecapTracker from 'Main/DeathRecapTracker';
+import SPELLS from 'common/SPELLS';
+
+class BloodDeathRecapTracker extends DeathRecapTracker {
+
+  on_initialized() {
+    this.defensiveBuffs = [
+      SPELLS.BONE_SHIELD.id,
+      SPELLS.DANCING_RUNE_WEAPON.id,
+    ];
+  }
+}
+
+export default BloodDeathRecapTracker;
