+++ src/common/DEFENSIVE_BUFFS.js
@@ -1,0 +1,15 @@
+import SPELLS from 'common/SPELLS';
+
+const DEFENSIVE_BUFFS = [
+  SPELLS.IRONBARK.id,
+  SPELLS.LIFE_COCOON.id,
+  SPELLS.BLESSING_OF_PROTECTION.id,
+  SPELLS.BLESSING_OF_SACRIFICE.id,
+  SPELLS.GUARDIAN_SPIRIT.id,
+  SPELLS.PAIN_SUPPRESSION.id,
+  SPELLS.POWER_WORD_BARRIER_BUFF.id,
+  //Blood Death-Knight
+  SPELLS.BONE_SHIELD.id,
+];
+
+export default DEFENSIVE_BUFFS;
